+++ src/libs/classNameCache.ts
@@ -247,10 +247,16 @@ export default class ClassNameCache {
         return;
       }
 
+      const line = Math.min(ruleStart.line - 1, document.lineCount - 1);
+      const character = Math.min(
+        ruleStart.column - 1,
+        document.lineAt(line).text.length
+      );
+
       if (
         await isPositionInComment(
           document,
-          new vscode.Position(ruleStart.line - 1, ruleStart.column - 1)
+          new vscode.Position(line, character)
         )
       ) {
         return;
@@ -262,9 +268,8 @@ export default class ClassNameCache {
 
           // Compute position in the document
           const selectorOffsetInDoc =
-            document.offsetAt(
-              new vscode.Position(ruleStart.line - 1, ruleStart.column - 1)
-            ) + sourceIndex;
+            document.offsetAt(new vscode.Position(line, character)) +
+            sourceIndex;
 
           const data: ClassNameRange = {
             range: new vscode.Range(