+++ CHANGELOG.md
@@ -4,6 +4,10 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+### Fixed
+
+- Minor Bugs
+
 ## [0.1.5] – 2026-01-27
 
 ### Added
@@ -135,5 +139,4 @@ All notable changes to this project will be documented in this file.
 [0.1.2]: https://github.com/Lokesh-Garg-22/CSS-Modules-IntelliSense/compare/v0.1.1...v0.1.2
 [0.1.3]: https://github.com/Lokesh-Garg-22/CSS-Modules-IntelliSense/compare/v0.1.2...v0.1.3
 [0.1.4]: https://github.com/Lokesh-Garg-22/CSS-Modules-IntelliSense/compare/v0.1.3...v0.1.4
-
 [0.1.5]: https://github.com/Lokesh-Garg-22/CSS-Modules-IntelliSense/compare/v0.1.4...v0.1.5